Extract SQLite merge implementation with resource cleanup
